Moved test data to it's own file
